Make the special generators compatible with static analysis
As those generators are re-exposing the Generator API through magic
methods, this makes static analyzers know about it thanks to the mixin
annotation.
Prior to 1.15, the generator methods were documented as returning
Generator instead of those special generators to make methods available,
but that was lying about the actual type (which would break typehints).