Change ii_IT text to use a public domain source
The poem "Pinocchio" by Collodi was referencing a CC-BY-SA-NC license
which prohibits commercial use. This may be a legal issue for software
that uses Faker.

As plenty other Text providers also use sources from the Gutenberg Project,
it seemed like a good idea to stick to that source as well for the italian
text.

When choosing an appropriate text, I had two objectives: it should not be
longer than the replaced text, and I looked for authors who lived before
the past 100 years to ensure that Copyright term has already passed.