Fix standard en_GB VAT numbers being generated with an incorrect number of digits
Standard UK VAT numbers should be 9 digits.

If the Modulus 97 block of a UK VAT number has a leading 0, this
zero gets thrown out when sprintf casts this block from a
string (%s) to an int (%d) during formatting.

For example, a check block of "07" would end up like:

GB216 5727 7

but it should be:

GB216 5727 07